+++ validations/book-validation.go
@@ -4,11 +4,11 @@ type BookUpdateValidation struct {
 	ID          uint64 `json:"id" form:"id" binding:"required"`
 	Title       string `json:"title" form:"title" binding:"required"`
 	Description string `json:"description" form:"description" binding:"required"`
-	UserID      uint16 `json:"user_id,omitempty" form:"user_id,omitempty"`
+	UserID      uint64 `json:"user_id,omitempty" form:"user_id,omitempty"`
 }
 
 type BookCreateValidation struct {
 	Title       string `json:"title" form:"title" binding:"required"`
 	Description string `json:"description" form:"description" binding:"required"`
-	UserID      uint16 `json:"user_id,omitempty" form:"user_id,omitempty"`
+	UserID      uint64 `json:"user_id,omitempty" form:"user_id,omitempty"`
 }
